requirements.txt
animeface
discord.py 
opencv-python
pillow 
requests
praw
# matplotlib 
# torch
# torchvision